// src/graphql/graphql-interfaces/dtos/resourceCompleted.interface.ts
import { CategoryDetailsCollectionDto } from "./categoryDetailsCollection.interface";
import { ScoreDetailsCollectionDto } from "./scoreDetailsCollection.interface";

export interface ResourceCompletedDto {
  id: string;
  tenantId: string;
  testId: string;
  firstName: string;
  lastName: string;
  studentGroupId: number;
  studentGroupName: string;
  studentGroup: string;
  testType: string;
  mappedTestType: string;
  sortField: string;
  testTitle: string;
  testName: string;
  testStartedAt: string;
  testCompletedAt: Date;
  testStatus: string;
  score: number;
  maxScore: number;
  scaledScore?: number;
  noOfQuestions: number;
  totalAnswered: number;
  normedCorrectPercentage: number;
  scoreThresholdPercentage: number;
  percentileRank: number;
  timeUsed: string;
  totalRemediationTime: number;
  timeSpent: string;
  navigationAlerts: number;
  percentCorrect: number;
  timesTaken: number;
  syllabusId: string;
  attemptNumber: number;
  avgRemediationTime: number;
  totalRemedQuestions: number;
  totalReviewTime: number;
  totalReviewedQuestions: number;
  numberOfAttempts?: number;
  scoreTo?: number;
  scoreFrom?: number;
  userId: string;
  atomEnrollmentId: string;
  classId: string;
  customGroupIds: string[];
  categoryDetails: CategoryDetailsCollectionDto[];
  scoreDetails: ScoreDetailsCollectionDto;
}